+++ app/src/main/res/layout/info_card.xml
@@ -9,7 +9,7 @@
         android:layout_gravity="center"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="2dp"
+        android:layout_marginBottom="4dp"
         android:layout_marginStart="3dp"
         android:layout_marginEnd="3dp"
         card_view:contentPaddingLeft="5dp"
@@ -47,8 +47,20 @@
 
 
             </LinearLayout>
+
+
+
         </LinearLayout>
 
+        <ImageView
+            android:layout_gravity="right"
+            android:id="@+id/linkImage"
+            android:src="@drawable/icon"
+            android:layout_height="wrap_content"
+            android:layout_width="wrap_content"
+            >
+
+        </ImageView>
 
     </android.support.v7.widget.CardView>
