--- app/src/main/res/layout/info_item.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent" android:layout_height="fill_parent"
-    android:padding="1dp"
-    android:background="@color/ugent_white">
-
-
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-
-        <RelativeLayout
-            android:layout_height="wrap_content"
-            android:layout_width="fill_parent"
-            android:id="@+id/headerLinearLay"
-            android:orientation="horizontal">
-            <ImageView
-                android:id="@+id/infoImage"
-                android:src="@drawable/icon"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"></ImageView>
-            <TextView
-                android:id="@+id/title"
-                android:text="Title"
-                android:textAppearance="?android:attr/textAppearanceLarge"
-                android:paddingLeft="8dp"
-                android:layout_toRightOf="@id/infoImage"
-                android:layout_centerVertical="true"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_vertical"
-                android:textColor="@color/ugent_blue_dark"></TextView>
-        </RelativeLayout>
-    </LinearLayout>
-
-</LinearLayout>
